backend/requirements.txt
# Plataforma DataScience Pro - Dependências Python

# FastAPI e ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de Dados
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
redis==5.0.1

# Machine Learning e Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
statsmodels==0.14.0
tensorflow==2.13.0
torch==2.0.1
xgboost==1.7.6
lightgbm==4.1.0
catboost==1.2.2

# AutoML
auto-sklearn==0.15.0
tpot==0.12.0
h2o==3.42.0.4

# Visualização
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
bokeh==3.2.2

# Processamento de Dados
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==13.0.0
polars==0.19.3

# APIs e Integração
requests==2.31.0
httpx==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Estatística Avançada
pingouin==0.5.3
lifelines==0.27.7
pymc==5.7.2
arviz==0.16.1

# Geração de Relatórios
jinja2==3.1.2
reportlab==4.0.4
fpdf2==2.7.6
python-docx==0.8.11

# Segurança e Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitários
python-dotenv==1.0.0
celery==5.3.4
redis-py-cluster==2.1.3
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
isort==5.12.0
mypy==1.6.1
